+++ src/main/java/com/barion/dungeons_enhanced/registry/DELootTables.java
@@ -4,15 +4,24 @@ import com.barion.dungeons_enhanced.DEUtil;
 import net.minecraft.resources.ResourceLocation;
 
 public interface DELootTables {
-    ResourceLocation SUNKEN_SHRINE = chest("sunken_shrine");
+    // Overworld
+    interface DESERT_TEMPLE {
+        ResourceLocation COFFIN = chest("desert_temple/coffin");
+        ResourceLocation FLOOR = chest("desert_temple/floor");
+        ResourceLocation TREASURE = chest("desert_temple/treasure");
+    }
+
     ResourceLocation MINERS_HOUSE = chest("miners_house");
 
-    interface PILLAGER_CAMP{
+    interface PILLAGER_CAMP {
         ResourceLocation GENERAL = chest("pillager_camp/general");
         ResourceLocation KITCHEN = chest("pillager_camp/kitchen");
     }
 
-    interface BLACK_CITADEL{
+    ResourceLocation SUNKEN_SHRINE = chest("sunken_shrine");
+
+    // Nether
+    interface BLACK_CITADEL {
         ResourceLocation NORMAL = chest("black_citadel/normal");
         ResourceLocation NORMAL_ALT = chest("black_citadel/normal_alt");
         ResourceLocation KITCHEN = chest("black_citadel/kitchen");
@@ -21,6 +30,6 @@ public interface DELootTables {
     }
 
     private static ResourceLocation chest(String key) {
-        return DEUtil.location("chests/" + key);
+        return DEUtil.locate("chests/" + key);
     }
 }