+++ src/main/java/com/barion/dungeons_enhanced/registry/DEProcessorLists.java
@@ -9,7 +9,7 @@ import net.minecraft.world.level.levelgen.structure.templatesystem.StructureProc
 
 import java.util.List;
 
-public class DEProcessorLists {
+public final class DEProcessorLists {
     public static final RegistrarHandler<StructureProcessorList> HANDLER = RegistrarHandler.getOrCreate(Registries.PROCESSOR_LIST, DungeonsEnhanced.MOD_ID);
 
     public static final Registrar.Pointer<StructureProcessorList> AIR_TO_COBWEB = HANDLER.createPointer("air_to_cobweb", ()-> listOf(DEProcessors.AIR_TO_COBWEB_2));